Make hunk-based application more universal.
That way it can be used for discarding hunks as well.
